pyproject.toml
[project]
name = "dit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "datasets>=3.6.0",
    "diffusers>=0.33.1",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "wandb>=0.19.11",
]
